Deprecate old InstrumentedFunction and restructure
