--- nevergrad/functions/core.py
@@ -1,8 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-# pylint: disable=unused-import
-# ExperimentFunction should be moved here when new parametrization is merged
-from nevergrad.instrumentation.multivariables import ExperimentFunction as ExperimentFunction
