+++ nevergrad/functions/mlda/problems.py
@@ -12,7 +12,7 @@ from typing import Optional
 import numpy as np
 import scipy.spatial
 from nevergrad.common.typetools import ArrayLike
-from ..core import ExperimentFunction
+from ..base import ExperimentFunction
 from ... import instrumentation as inst
 from . import datasets
 
@@ -45,6 +45,7 @@ class Clustering(ExperimentFunction):
             self._points -= np.mean(self._points, axis=0, keepdims=True)
             self._points /= np.std(self._points, axis=0, keepdims=True)
         super().__init__(self._compute_distance, inst.var.Array(num_clusters, points.shape[1]))
+        self.register_initialization(points=points, num_clusters=num_clusters, rescale=rescale)
         self._descriptors.update(num_clusters=num_clusters, rescale=rescale)
 
     @classmethod
@@ -94,6 +95,7 @@ class Perceptron(ExperimentFunction):
         self._x = x
         self._y = y
         super().__init__(self._compute_loss, inst.var.Array(10))
+        self.register_initialization(x=x, y=y)
 
     @classmethod
     def from_mlda(cls, name: str) -> "Perceptron":
@@ -151,6 +153,7 @@ class SammonMapping(ExperimentFunction):
         self._proximity_2 = self._proximity**2
         self._proximity_2[self._proximity_2 == 0] = 1  # avoid ZeroDivision (for diagonal terms, or identical points)
         super().__init__(self._compute_distance, inst.var.Array(self._proximity.shape[0], 2))
+        self.register_initialization(proximity_array=proximity_array)
 
     @classmethod
     def from_mlda(cls, name: str, rescale: bool = False) -> "SammonMapping":
@@ -227,6 +230,7 @@ class Landscape(ExperimentFunction):
 
     def __init__(self, transform: Optional[str] = None) -> None:
         super().__init__(self._get_pixel_value, inst.Instrumentation(inst.var.Scalar(), inst.var.Scalar()))
+        self.register_initialization(transform=transform)
         self.parametrization = self.parametrization.with_name("standard")  # force descriptor update
         self._image = datasets.get_data("Landscape")
         if transform == "gaussian":
