+++ nevergrad/functions/photonics/core.py
@@ -24,7 +24,7 @@ import shutil
 from typing import List
 from pathlib import Path
 import numpy as np
-from ..core import ExperimentFunction
+from ..base import ExperimentFunction
 from ... import instrumentation as inst
 
 
@@ -126,6 +126,7 @@ class Photonics(ExperimentFunction):
                                           cwd=path.parent, verbose=False,
                                           env=dict(os.environ, OMP_NUM_THREADS="1", OPENBLAS_NUM_THREADS="1"))
         super().__init__(self._compute, _make_instrumentation(name=name, dimension=dimension, transform=transform))
+        self.register_initialization(name=name, dimension=dimension, transform=transform)
         self._descriptors.update(name=name)
 
     def _compute(self, *x: np.ndarray) -> float:
