+++ nevergrad/functions/stsp/core.py
@@ -12,13 +12,14 @@
 import matplotlib.pyplot as plt
 import numpy as np
 from nevergrad import instrumentation as inst
-from ..core import ExperimentFunction
+from ..base import ExperimentFunction
 
 
 class STSP(ExperimentFunction):
 
     def __init__(self, dimension: int = 500) -> None:
         super().__init__(self._simulate_stsp, inst.var.Array(dimension))
+        self.register_initialization(dimension=dimension)
         self.order = np.arange(0, self.dimension)
         self.x = self.parametrization.random_state.normal(size=self.dimension)
         self.y = self.parametrization.random_state.normal(size=self.dimension)
@@ -31,9 +32,6 @@ class STSP(ExperimentFunction):
         output = np.sqrt((x[0] - x[-1])**2 + (y[0] - y[-1])**2) + sum(np.sqrt((x[i] - x[i + 1])**2 + (y[i] - y[i + 1])**2)
                                                                       for i in range(self.dimension - 1))
         return float(output)
-
-    def copy(self) -> "STSP":
-        return STSP(self.dimension)
 
     def make_plots(self, filename: str = "stsp.png") -> None:
         plt.clf()
