+++ nevergrad/instrumentation/multivariables.py
@@ -4,7 +4,6 @@
 # LICENSE file in the root directory of this source tree.
 
 import itertools
-import typing as tp
 from typing import List, Any, Tuple, Dict, Optional, Callable
 import numpy as np
 from ..common.typetools import ArrayLike
@@ -211,225 +210,15 @@ class Instrumentation(NestedVariables):
         return super()._arguments_to_data(*(wrap_arg(x) for x in args), **{k: wrap_arg(x) for k, x in kwargs.items()})
 
 
-Parameter = Instrumentation  # looking ahead
-PF = tp.TypeVar("PF", bound="ExperimentFunction")
-
-
-class ExperimentFunction:
-    """Combines a function and its parametrization for running experiments
-
-    Parameters
-    ----------
-    function: callable
-        the callable to convert
-    parametrization: Parameter
-        the parametrization of the function
-    Notes
-    -----
-    - This function can then be directly used in benchmarks *if it returns a float*.
-    - You can update the "_descriptors" dict attribute so that function parameterization is recorded during benchmark
-    """
-
-    def __init__(self, function: tp.Callable[..., float], parametrization: Variable) -> None:
-        assert callable(function)
-        self._descriptors: tp.Dict[str, tp.Any] = {"function_class": self.__class__.__name__}
-        self._parametrization = Instrumentation()
-        self.parametrization = parametrization if isinstance(parametrization, Instrumentation) else Instrumentation(parametrization)
-        self._function = function
-        # if this is not a function bound to this very instance, add the function/callable name to the descriptors
-        if not hasattr(function, '__self__') or function.__self__ != self:  # type: ignore
-            name = function.__name__ if hasattr(function, "__name__") else function.__class__.__name__
-            self._descriptors.update(name=name)
-        if hasattr(self, "get_postponing_delay"):
-            raise RuntimeError('"get_posponing_delay" has been replaced by "compute_pseudotime" and has been  aggressively deprecated')
-        if hasattr(self, "noisefree_function"):
-            raise RuntimeError('"noisefree_function" has been replaced by "evaluation_function" and has been  aggressively deprecated')
-
-    @property
-    def dimension(self) -> int:
-        return self._parametrization.dimension
-
-    @property
-    def parametrization(self) -> Parameter:
-        return self._parametrization
-
-    @parametrization.setter
-    def parametrization(self, parametrization: Parameter) -> None:
-        self._parametrization = parametrization
-        self._parametrization.freeze()
-        # TODO change to parametrization
-        self._descriptors.update(instrumentation=parametrization.name, dimension=parametrization.dimension)
-
-    @property
-    def function(self) -> tp.Callable[..., float]:
-        return self._function
-
-    def __call__(self, *args: tp.Any, **kwargs: tp.Any) -> float:
-        """Call the function directly (equivaluent to parametrized_function.function(*args, **kwargs))
-        """
-        return self._function(*args, **kwargs)
-
-    @property
-    def descriptors(self) -> tp.Dict[str, tp.Any]:
-        """Description of the function parameterization, as a dict. This base class implementation provides function_class,
-            noise_level, transform and dimension
-        """
-        return dict(self._descriptors)  # Avoid external modification
-
-    def __repr__(self) -> str:
-        """Shows the function name and its summary
-        """
-        params = [f"{x}={repr(y)}" for x, y in sorted(self._descriptors.items())]
-        return "Instance of {}({})".format(self.__class__.__name__, ", ".join(params))
-
-    def equivalent_to(self, other: tp.Any) -> bool:
-        """Check that two instances where initialized with same settings.
-        This is not meant to be used to check if functions are exactly equal
-        (initialization may hold some randomness)
-        This is only useful for unit testing.
-        (may need to be overloaded to make faster if tests are getting slow)
-        """
-        if other.__class__ != self.__class__:
-            return False
-        return bool(self._descriptors == other._descriptors)
-
-    def copy(self: PF) -> PF:
-        """Provides a new equivalent instance of the class, possibly with
-        different random initialization, to provide different equivalent test cases
-        when using different seeds.
-        """
-        pf = self.__class__(self.function, self.parametrization.copy())
-        pf._descriptors = self.descriptors
-        return pf
-
-    def compute_pseudotime(self, input_parameter: Any, value: float) -> float:  # pylint: disable=unused-argument
-        """Computes a pseudotime used during benchmarks for mocking parallelization in a reproducible way.
-        By default, each call takes 1 unit of pseudotime, but this can be modified by overriding this
-        function and the pseudo time can be a function of the function inputs and output.
-
-        Note: This replaces get_postponing_delay which has been aggressively deprecated
-
-        Parameters
-        ----------
-        input_parameter: Any
-            the input that was provided to the actual function
-        value: float
-            the output of the actual function
-
-        Returns
-        -------
-        float
-            the pseudo computation time of the call to the actual function
-        """
-        return 1.
-
-    def evaluation_function(self, *args: Any, **kwargs: Any) -> float:
-        """Provides a (usually "noisefree") function used at final test/evaluation time in benchmarks.
-
-        Parameters
-        ----------
-        *args, **kwargs
-            same as the actual function
-        """
-        return self.function(*args, **kwargs)
-
-
-class InstrumentedFunction(ExperimentFunction):
-    """Converts a multi-argument function into a single-argument multidimensional continuous function
-    which can be optimized.
-
-    Parameters
-    ----------
-    function: callable
-        the callable to convert
-    *args, **kwargs: Any
-        Any argument. Arguments of type Variable (see notes) will be instrumented,
-        and others will be kept constant.
-
-    Notes
-    -----
-    - Variable classes are:
-        - `SoftmaxCategorical(items)`: converts a list of `n` (unordered) categorial items into an `n`-dimensional space. The returned
-           element will be sampled as the softmax of the values on these dimensions. Be cautious: this process is non-deterministic
-           and makes the function evaluation noisy.
-        - `OrderedDiscrete(items)`: converts a list of (ordered) discrete items into a 1-dimensional variable. The returned value will
-           depend on the value on this dimension: low values corresponding to first elements of the list, and high values to the last.
-        - `Gaussian(mean, std)`: normalizes a `n`-dimensional variable with independent Gaussian priors (1-dimension per value).
-        - `Array(dim1, ...)`: casts the data from the optimization space into a np.ndarray of any shape,
-          to which some transforms can be applied (see `asscalar`, `affined`, `exponentiated`, `bounded`).
-          This is therefore a very flexible type of variable.
-        - `Scalar(dtype)`: casts the data from the optimization space into a float or an int. It is equivalent to `Array(1).asscalar(dtype)`
-          and all `Array` methods are therefore available
-    - This function can then be directly used in benchmarks *if it returns a float*.
-    - You can update the "_descriptors" dict attribute so that function parameterization is recorded during benchmark
+class InstrumentedFunction:
+    """InstrumentedFunction is being aggressively deprecated. Conversion depends on your use case:
+    - for optimization purpose: directly provide ng.Instrumentation to the optimizer, it will
+       provide candidates with fields 'args' and 'kwargs' that match the instrumentation.
+    - for benchmarks: derive from ng.functions.ExperimentFunction. Main differences are:
+       calls to __call__ directly forwards the main function (instead of converting from data space),
+       __init__ takes exactly two arguments (main function and parametrization/instrumentation) and
+      instrumentation is renamed to parametrization for forward compatibility.
     """
 
     def __init__(self, function: Callable[..., Any], *args: Any, **kwargs: Any) -> None:
-        assert callable(function)
-        self._descriptors: Dict[str, Any] = {"function_class": self.__class__.__name__}
-        super().__init__(function, Instrumentation(*args, **kwargs))
-        self.last_call_args: Optional[Tuple[Any, ...]] = None
-        self.last_call_kwargs: Optional[Dict[str, Any]] = None
-
-    @property
-    def instrumentation(self) -> Instrumentation:
-        return self.parametrization
-
-    @instrumentation.setter
-    def instrumentation(self, instrum: Instrumentation) -> None:
-        self.parametrization = instrum
-
-    @property
-    def dimension(self) -> int:
-        return self.instrumentation.dimension
-
-    def data_to_arguments(self, data: ArrayLike, deterministic: bool = True) -> Tuple[Tuple[Any, ...], Dict[str, Any]]:
-        """Get the arguments and keyword arguments corresponding to the data
-
-        Parameters
-        ----------
-        data: np.ndarray
-            input data
-        deterministic: bool
-            whether to process the data deterministically (some Variables such as SoftmaxCategorical are stochastic).
-            If True, the output is the most likely output.
-        """
-        return self.instrumentation.data_to_arguments(data, deterministic=deterministic)
-
-    def legacy_call(self, x: ArrayLike) -> Any:
-        self.last_call_args, self.last_call_kwargs = self.data_to_arguments(x, deterministic=False)
-        return self.function(*self.last_call_args, **self.last_call_kwargs)
-
-    def arguments_to_data(self, *args: Any, **kwargs: Any) -> ArrayLike:
-        return self.instrumentation.arguments_to_data(*args, **kwargs)
-
-    def get_summary(self, data: ArrayLike) -> Any:  # probably impractical for large arrays
-        """Provides the summary corresponding to the provided data
-        """
-        return self.instrumentation.get_summary(data)
-
-    @property
-    def descriptors(self) -> Dict[str, Any]:
-        """Description of the function parameterization, as a dict. This base class implementation provides function_class,
-            noise_level, transform and dimension
-        """
-        return dict(self._descriptors)  # Avoid external modification
-
-    def __repr__(self) -> str:
-        """Shows the function name and its summary
-        """
-        params = [f"{x}={repr(y)}" for x, y in sorted(self._descriptors.items())]
-        return "Instance of {}({})".format(self.__class__.__name__, ", ".join(params))
-
-    def __eq__(self, other: Any) -> bool:
-        """Check that two instances where initialized with same settings.
-        This is not meant to be used to check if functions are exactly equal (initialization may hold some randomness)
-        This is only useful for unit testing.
-        (may need to be overloaded to make faster if tests are getting slow)
-        """
-        if other.__class__ != self.__class__:
-            return False
-        return bool(self._descriptors == other._descriptors)
-
-    def copy(self: PF) -> PF:
-        raise RuntimeError("InstrumentedFunction does not implement copy")
+        raise RuntimeError(self.__doc__)
